--- src/main/java/AppByDataBaseStorageHibernate.java
@@ -1,11 +1,0 @@
-import runner.DataBaseHibernateModeRunner;
-
-public class AppByDataBaseStorageHibernate {
-    /**
-     * Method description
-     *
-     */
-    static void start() {
-        new DataBaseHibernateModeRunner().start();
-    }
-}
