+++ src/main/java/config/DbType.java
@@ -1,5 +1,5 @@
 package config;
 
 public enum DbType {
-    STORAGE_JSON, STORAGE_DATABASE_JDBC, STORAGE_DATABASE_HIBERNATE
+    STORAGE_JSON, STORAGE_DATABASE
 }
