+++ src/main/resources/config.properties
@@ -1,11 +1,12 @@
 
-#STORAGE_JSON or STORAGE_DATABASE_JDBC or STORAGE_DATABASE_HIBERNATE
+#STORAGE_JSON or STORAGE_DATABASE or STORAGE_DATABASE_HIBERNATE
 storage_type=STORAGE_DATABASE_HIBERNATE
 
 
 #if storage_type is STORAGE_DATABASE or STORAGE_DATABASE_HIBERNATE you must connect your database
 #if you don't have table storage in your db use create_table.sql file
-
+#choose connection type to db: by_jdbc or by_hibernate
+db.connection_type=by_hibernate
 db.url=jdbc:postgresql://localhost:5432/storage_database
 db.username=postgres
 db.password=riviera0685
